feat: clubname 변경 기능 추가

+++ projectDongbi/src/main/java/com/dongbi/projectDongbi/domain/club/repository/ClubRepository.java
@@ -3,5 +3,5 @@ package com.dongbi.projectDongbi.domain.club.repository;
 import com.dongbi.projectDongbi.domain.club.Club;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClubRepository extends CrudRepository<Club, Integer> {
+public interface ClubRepository extends CrudRepository<Club, Long> {
 }
